LPS-119021 Verify if DDMContent exists before remove it

+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-service/src/main/java/com/liferay/dynamic/data/mapping/internal/storage/JSONDDMStorageAdapter.java
@@ -54,8 +54,13 @@ public class JSONDDMStorageAdapter implements DDMStorageAdapter {
 		throws StorageException {
 
 		try {
-			ddmContentLocalService.deleteDDMContent(
+			DDMContent ddmContent = ddmContentLocalService.fetchDDMContent(
 				ddmStorageAdapterDeleteRequest.getPrimaryKey());
+
+			if (ddmContent != null) {
+				ddmContentLocalService.deleteDDMContent(
+					ddmContent.getPrimaryKey());
+			}
 
 			DDMStorageAdapterDeleteResponse.Builder builder =
 				DDMStorageAdapterDeleteResponse.Builder.newBuilder();
